backend/app/routers/users.py
from fastapi import Depends, HTTPException, Session
from jose import JWTError, jwt
from app.models import User
from app.database import get_db
from fastapi.security import OAuth2PasswordBearer

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="login")

# Секретный ключ для расшифровки токенов
SECRET_KEY = "your_secret_key"
ALGORITHM = "HS256"

# Функция для извлечения пользователя из токена
def get_current_user(token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)) -> User:
    try:
        # Декодируем токен и получаем информацию о пользователе
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username: str = payload.get("sub")
        if username is None:
            raise HTTPException(status_code=401, detail="Invalid token")

        # Находим пользователя в базе данных
        user = db.query(User).filter(User.username == username).first()
        if user is None:
            raise HTTPException(status_code=404, detail="User not found")
        return user
    except JWTError:
        raise HTTPException(status_code=401, detail="Could not validate credentials")
